any element can be accessed by event.target.root.getElementById(<element's id>)
all ids must be unique and the library enforces the constraint.

by default all ids are null which is not problematic. when an id is specified, it is checked against all other elements. If any two elements with duplicate ids are found anywhere, the latter one is not added to preserve the unique id constraint.